+++ src/commands/mod/ban.js
@@ -1,5 +1,5 @@
 /**
- * Discord Welcome bot
+ * Discord Welcome-Bot
  * Copyright (c) 2021 The Welcome-Bot Team and Contributors
  * Licensed under Lesser General Public License v2.1 (LGPl-2.1 - https://opensource.org/licenses/lgpl-2.1.php)
  */
@@ -58,9 +58,9 @@ module.exports = class CMD extends Command {
             return message.channel.send(`Failed to ban **${user.tag}**`);
         }
 
-        if (guildDB.modChannel) {
+        if (guildDB.plugins.modlogs) {
             channel = message.guild.channels.cache.find(
-                (ch) => ch.name === guildDB.modChannel
+                (ch) => ch.name === guildDB.plugins.modlogs
             );
             if (channel) {
                 const embed = new Embed({ color: "red" });